Draft to support JUnit's extension model
Benchmark classes and methods can now be annotated with conditional annotations such as Disabled or EnabledIf… for conditional benchmark execution. Annotations are evaluated against the ExtensionRegistry and are extensible by providing a ExecutionCondition extension.

Annotations are evaluated within the starting JVM to determine which benchmarks to run and which to not include in the run. Any state changes (e.g. changes to static fields) are not visible in the actual benchmark run as JMH has its own JVM runtime lifecycle management.